refractor: move memory router up one level
