+++ src/App.tsx
@@ -8,7 +8,7 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { ThemeProvider } from '@emotion/react';
 import { CssBaseline } from '@mui/material';
 
-import { TransactionList } from './pages/Transaction';
+import Transaction from './pages/Transaction';
 import { theme } from './theme';
 
 export function App() {
@@ -18,7 +18,7 @@ export function App() {
       <Router>
         <Switch>
           <Route path="/">
-            <TransactionList />
+            <Transaction />
           </Route>
         </Switch>
       </Router>
